requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.2.2
Flask==3.1.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-documentai==3.2.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymongo==4.11.2
pypdf==4.1.0
pytest==8.3.5
python-dotenv==1.0.1
reportlab==4.3.1
requests==2.32.3
rsa==4.9
tomli==2.2.1
urllib3>=2.0.7
Werkzeug==3.1.3
chardet==5.2.0
